+++ chancy/plugins/api/__init__.py
@@ -10,7 +10,7 @@ from starlette.middleware.cors import CORSMiddleware
 from starlette.staticfiles import StaticFiles
 
 from chancy import Worker, Chancy
-from chancy.plugin import Plugin, PluginScope
+from chancy.plugin import Plugin
 from chancy.plugins.api.core import CoreApiPlugin
 from chancy.plugins.api.plugin import ApiPlugin
 from chancy.utils import import_string
@@ -63,6 +63,18 @@ class Api(Plugin):
         public internet. It is intended for use in a secure environment, such
         as a private network or a VPN where only trusted users have access.
 
+    Since it's very common to only want the dashboard temporarily, you can
+    start it with the CLI. This can also be used to connect to a remote Chancy
+    instance:
+
+    .. code-block:: bash
+
+        pip install chancy[cli,web]
+        chancy --app worker.chancy worker web
+
+    This will run the API and dashboard on port 8000 by default (you can change
+    this with the ``--port`` and ``--host`` flags).
+
     Screenshots
     -----------
 
@@ -79,29 +91,11 @@ class Api(Plugin):
         :alt: Worker page
 
 
-    CLI
-    ---
-
-    Since it's very common to only want the dashboard temporarily, you can
-    start it with the CLI:
-
-    .. code-block:: bash
-
-        pip install chancy[cli,web]
-        chancy --app worker.chancy worker web
-
-    This will run the API and dashboard on port 8000 by default (you can change
-    this with the ``--port`` and ``--host`` flags).
-
     :param port: The port to listen on.
     :param host: The host to listen on.
     :param debug: Whether to run the server in debug mode.
     :param allow_origins: A list of origins that are allowed to access the API.
     """
-
-    @classmethod
-    def get_scope(cls) -> PluginScope:
-        return PluginScope.WORKER
 
     def __init__(
         self,
@@ -119,11 +113,15 @@ class Api(Plugin):
         self.allow_origins = allow_origins or []
         self.plugins: set[Type[ApiPlugin]] = {CoreApiPlugin}
 
+    @staticmethod
+    def get_identifier() -> str:
+        return "chancy.api"
+
     async def run(self, worker: Worker, chancy: Chancy):
         """
         Start the web server.
         """
-        for plug in chancy.plugins:
+        for plug in chancy.plugins.values():
             api_plugin = plug.api_plugin()
             if api_plugin is None:
                 continue